+++ CollectorHelper.lua
@@ -1,217 +1,41 @@
-local itemIndexMap = {}
-local forceShowMerchant = false
+local _, app = ...
+
 local merchantIndex = {}
+local sourceHidden = true
+local COLORS = app.COLORS
 
 SLASH_COLLECTORHELPER1 = "/ch"
 SlashCmdList["COLLECTORHELPER"] = function()
-  settings.hideMerchantOwned = not settings.hideMerchantOwned
-  updateShop()
+  -- do nothing
 end
 
--- ========================
--- Section: Merchant Frame build
--- ========================
-local merchantFrameCost = frameBuilder({
-  frameName = "CollectorHelper_MerchantButton",
-  parent = MerchantFrame,
-  width = 280,
-  height = MerchantFrame:GetHeight(),
-  point = {
-    pos = "TOPRIGHT",
-    x = 280,
-    y = 0,
-  },
-  titleBuilder = {
-    text = textCFormat(COLORS.yellow, "Currency Needed to Collect this page"),
-    point = {
-      pos = "TOP",
-      x = 0,
-      y = -8,
-    }
-  }
-})
-
--- cost details
-local marchantCost = fontBuilder({
-  parent = merchantFrameCost,
-  text = "",
-  point = {
-    pos = "TOPLEFT",
-    x = 35,
-    y = -40,
-  }
-})
-
-local m1s = buttonBuilder({
-  buttonName = "Collector_MBuyButton",
-  parent = merchantFrameCost,
-  text = "Buy All Possible",
-  width = 120,
-  height = 22,
-  point = {
-    pos = "BOTTOMLEFT",
-    x = 18,
-    y = 8,
-  }
-})
-
-m1s:HookScript("OnClick", function(self, button)
-  if button == "LeftButton" then
-    for itemIndex, itemId in pairs(itemIndexMap) do
-      local costInfo = GetMerchantItemCostInfo(itemIndex)
-      if costInfo == 0 or CanAffordMerchantItem(itemIndex) then
-        BuyMerchantItem(itemIndex, 1)
-      end
-    end
-  end
-end)
-
-local m2s = buttonBuilder({
-  buttonName = "Collector_MSHButton",
-  parent = merchantFrameCost,
-  text = "Show/hide owned",
-  width = 120,
-  height = 22,
-  point = {
-    pos = "BOTTOMRIGHT",
-    x = -18,
-    y = 8,
-  }
-})
-
-m2s:HookScript("OnClick", function(self, button)
-  if button == "LeftButton" then
-    settings.hideMerchantOwned = not settings.hideMerchantOwned
-    if settings.hideMerchantOwned == false then
-      forceShowMerchant = true
-    end
-    updateShop()
-  end
-end)
-
-
--- ========================
--- Section: Item owned logic
--- ========================
-local function isToyItem(itemId)
-  if C_ToyBox.GetToyInfo(itemId) then
-    return true
-  end
-  return false
-end
-
-local function IsMountItem(itemId)
-  if C_MountJournal.GetMountFromItem(itemId) then
-    return true
-  end
-  return false
-end
-
-local function IsPetItem(itemId)
-  if C_PetJournal.GetPetInfoByItemID(itemId) then
-    return true
-  end
-  return false
-end
-
--- check set owned
-local function setOwned(itemSet)
-  local ids = C_TransmogSets.GetAllSourceIDs(itemSet)
-  for _, sourceID in ipairs(ids) do
-    local info = C_TransmogCollection.GetSourceInfo(sourceID)
-    if not info.isCollected then return false end
-  end
-  return true
-end
-
--- check item owned from merchant
-local function checkShopID(source)
-  -- 0 false, 1 true, 2 ignore
-  if source == nil or source.itemId == nil then
-    return 0
-  end
-
-  local isToy = isToyItem(source.itemId)
-  if isToy then
-    return PlayerHasToy(source.itemId) and 1 or 0
-  end
-
-  local isMount = IsMountItem(source.itemId)
-  if isMount then
-    local mountID = C_MountJournal.GetMountFromItem(source.itemId)
-    local playerKnowsMount = select(11, C_MountJournal.GetMountInfoByID(mountID))
-    return playerKnowsMount and 1 or 0
-  end
-
-  local IsPet = IsPetItem(source.itemId)
-  if IsPet then
-    local petName = C_PetJournal.GetPetInfoByItemID(source.itemId)
-    if petName ~= nil then
-      local _, petGUID = C_PetJournal.FindPetIDByName(petName)
-      return petGUID ~= nil and 1 or 0
-    end
-  end
-
-  local itemSetId = C_Item.GetItemLearnTransmogSet(source.itemId)
-  local isItemHeirloom = C_Heirloom.IsItemHeirloom(source.itemId)
-  if itemSetId ~= nil then
-    return setOwned(itemSetId) and 1 or 0
-  elseif isItemHeirloom then
-    return C_Heirloom.PlayerHasHeirloom(source.itemId) and 1 or 0
-  else
-    local itemAppearanceID, _ = C_TransmogCollection.GetItemInfo(source.itemId)
-    if itemAppearanceID == nil then
-      local sourceID = GetSourceID(source.link)
-      if sourceID ~= nil then
-        return C_TransmogCollection.PlayerHasTransmogItemModifiedAppearance(sourceID) and 1 or 0
-      else
-        return 2 -- ignore item
-      end
-    else
-      local r = C_TransmogCollection.PlayerHasTransmog(source.itemId) and 1 or 0
-      if r == 0 and source.bindType == 2 then
-        if playerHasItemInBag(source.itemId) == true then
-          return 10
-        end
-      end
-      return r
-    end
-  end
-end
-
-function playerHasItemInBag(itemCheckId)
-  -- check if item is in bag
-  for bag = 0, NUM_BAG_SLOTS do
-    for slot = 1, C_Container.GetContainerNumSlots(bag) do
-      local item = C_Container.GetContainerItemInfo(bag, slot)
-      if (item ~= nil and item.itemID == itemCheckId) then
-        return true
-      end
-    end
-  end
-  return false
-end
 
 -- ========================
 -- Section: Merchant items
 -- ========================
-function updateShop()
+function app:updateShop()
   local itemData = {}
-  local size = MERCHANT_ITEMS_PER_PAGE
+  local size = MERCHANT_ITEMS_PER_PAGE ---@type number
   local currencyMap = {}
+  local currencyLink = {}
   local initialItemIndexMap = {}
-  merchantIndex = {}
+  merchantIndex = {}        -- reset
+  app.merchantTokenMap = {} -- reset
 
   for i = 1, size do
-    local itemIndex = (MerchantFrame.page - 1) * size + i
-    local itemId = GetMerchantItemID(itemIndex)
+    local itemIndex = (MerchantFrame.page - 1) * size + i ---@type number
+    local itemId = GetMerchantItemID(itemIndex) ---@type number|nil
     if itemId ~= nil then
       merchantIndex[i] = itemId
-      local source = getItemDetails(itemId)
-      local shopItemState = checkShopID(source)
-      local equipBtn = _G["MerchantItem" .. i .. "EquipBtn"]
+      local source = app:getItemDetails(itemId)
+      local shopItemState = app:checkShopID(source)
+      local equipBtn = _G["MerchantItem" .. i .. "EquipBtn"] ---@type CH.Btn
+      local showSources = _G["MerchantItem" .. i .. "ShowSrcBtn"] ---@type CH.Btn
       if equipBtn ~= nil then
         equipBtn:Hide()
+      end
+      if showSources ~= nil then
+        showSources:Hide()
       end
       if shopItemState == 1 then
         if settings.hideMerchantOwned then
@@ -220,7 +44,7 @@ function updateShop()
           _G["MerchantItem" .. i .. "Name"]:SetText("")
           _G["MerchantItem" .. i .. "AltCurrencyFrame"]:Hide()
           _G["MerchantItem" .. i .. "MoneyFrame"]:Hide()
-        elseif forceShowMerchant == true then
+        elseif app.forceShowMerchant == true then
           local currencyIndex = GetMerchantItemCostInfo(itemIndex)
           SetItemButtonSlotVertexColor(_G["MerchantItem" .. i], 1, 1, 1)
           _G["MerchantItem" .. i .. "ItemButton"]:Show()
@@ -231,36 +55,60 @@ function updateShop()
             _G["MerchantItem" .. i .. "AltCurrencyFrame"]:Show()
           end
         end
-      elseif shopItemState == 0 then
+      elseif shopItemState == 0 or shopItemState == 15 then
         local currencyIndex = GetMerchantItemCostInfo(itemIndex)
+        local itemsSources = 1
+        if shopItemState == 15 then -- TODO items with sources
+          local notOwnedCount = 0
+          for _, entry in ipairs(app.merchantTokenMap[source.itemId]) do
+            if not entry.owned then
+              notOwnedCount = notOwnedCount + 1
+            end
+          end
+          local srcBtn = _G["MerchantItem" .. i .. "ShowSrcBtn"] ---@type CH.Btn
+          if srcBtn == nil then
+            handleSourceMerchantToken(i)
+          else
+            srcBtn:Show()
+          end
+          itemsSources = notOwnedCount
+        end
         if currencyIndex == 0 then
           -- is gold
           local _, _, price, _, _, _, _, _ = GetMerchantItemInfo(itemIndex)
-          local itemTexture = "MoneyCurrency"
-          if itemTexture then
-            if currencyMap[itemTexture] then
-              currencyMap[itemTexture] = currencyMap[itemTexture] + price
-            else
-              currencyMap[itemTexture] = price
+          for _ = 1, itemsSources do
+            local itemTexture = "MoneyCurrency"
+            if itemTexture then
+              currencyLink[itemTexture] = nil
+              if currencyMap[itemTexture] then
+                currencyMap[itemTexture] = currencyMap[itemTexture] + price
+              else
+                currencyMap[itemTexture] = price
+              end
+              local count = GetMoney()
+              itemData[itemTexture] = count
             end
-            local count = GetMoney()
-            itemData[itemTexture] = count
           end
         else
           for y = 1, currencyIndex do
-            local itemTexture, itemValue, link, _ = GetMerchantItemCostItem(itemIndex, y)
-            if itemTexture then
-              if currencyMap[itemTexture] then
-                currencyMap[itemTexture] = currencyMap[itemTexture] + itemValue
-              else
-                currencyMap[itemTexture] = itemValue
-              end
-              local ci = C_CurrencyInfo.GetCurrencyInfoFromLink(link)
-              if ci ~= nil then
-                itemData[itemTexture] = ci.quantity
-              else
-                local count = C_Item.GetItemCount(link)
-                itemData[itemTexture] = count
+            local itemTexture, itemValue, link = GetMerchantItemCostItem(itemIndex, y)
+            for _ = 1, itemsSources do
+              if itemTexture then
+                if currencyLink[itemTexture] == nil then
+                  currencyLink[itemTexture] = link
+                end
+                if currencyMap[itemTexture] then
+                  currencyMap[itemTexture] = currencyMap[itemTexture] + itemValue
+                else
+                  currencyMap[itemTexture] = itemValue
+                end
+                local ci = C_CurrencyInfo.GetCurrencyInfoFromLink(link)
+                if ci ~= nil then
+                  itemData[itemTexture] = ci.quantity
+                else
+                  local count = C_Item.GetItemCount(link)
+                  itemData[itemTexture] = count
+                end
               end
             end
           end
@@ -271,9 +119,9 @@ function updateShop()
         _G["MerchantItem" .. i .. "Name"]:SetText("This item is in your bag")
         _G["MerchantItem" .. i .. "AltCurrencyFrame"]:Hide()
         _G["MerchantItem" .. i .. "MoneyFrame"]:Hide()
-        local eqBtn = _G["MerchantItem" .. i .. "EquipBtn"]
+        local eqBtn = _G["MerchantItem" .. i .. "EquipBtn"] ---@type CH.Btn
         if eqBtn == nil then
-          merchantEquipHandler(i)
+          app:merchantEquipHandler(i)
         else
           eqBtn:Show()
         end
@@ -282,16 +130,20 @@ function updateShop()
       end
     end
   end
-  forceShowMerchant = false
+  app.forceShowMerchant = false
 
-  local cost = ""
+  local costTable = {}
+
+  local missingItem = false
   for itemTexture, totalValue in pairs(currencyMap) do
+    local display = ""
+    local description = ""
     local haveCurencyVal = itemData[itemTexture]
     local isGold = itemTexture == "MoneyCurrency"
     if isGold then
-      cost = cost .. formatGoldAmount(GetMoneyString(totalValue, true))
+      display = display .. app:formatGoldAmount(GetMoneyString(totalValue, true))
     else
-      cost = cost .. "|T" .. itemTexture .. ":16|t " .. totalValue
+      display = display .. "|T" .. itemTexture .. ":16|t " .. totalValue
     end
     if haveCurencyVal ~= nil then
       local percentage = (haveCurencyVal / totalValue) * 100
@@ -299,19 +151,30 @@ function updateShop()
         percentage = 100
       end
       if isGold then
-        haveCurencyVal = formatGoldAmount(GetMoneyString(haveCurencyVal, true))
+        haveCurencyVal = app:formatGoldAmount(GetMoneyString(haveCurencyVal, true))
       end
-      cost = cost .. " - " ..
-          textCFormat(COLORS.green, haveCurencyVal) .. " (" .. string.format("%.2f", percentage) .. "%)"
+      display = display .. " - " ..
+          app:textCFormat(COLORS.green, haveCurencyVal) .. " (" .. string.format("%.2f", percentage) .. "%)"
     end
-    cost = cost .. "\n\n"
+
+    table.insert(costTable, {
+      display = display,
+      description = description,
+      isGold = isGold,
+      linkItem = currencyLink[itemTexture]
+    })
+    missingItem = true
   end
 
-  if cost == "" then
-    cost = textCFormat(COLORS.green, "You have everything on this page")
+  if missingItem == true then
+    app.marchantCost:Hide()
+    app.mainScrollableContent.scrollFrame:Show()
+    app.mainScrollableContent.UpdateRows(costTable)
+  else
+    app.marchantCost:Show()
+    app.mainScrollableContent.scrollFrame:Hide()
+    app.mainScrollableContent.UpdateRows(costTable)
   end
-  marchantCost:SetText(cost)
-
   -- Create a sorted list of itemIndex keys
   local sortedItemIndices = {}
   for itemIndex in pairs(initialItemIndexMap) do
@@ -320,16 +183,20 @@ function updateShop()
   table.sort(sortedItemIndices)
 
   -- Create a new sorted map
-  itemIndexMap = {}
+  app.itemIndexMap = {}
   for _, itemIndex in ipairs(sortedItemIndices) do
-    itemIndexMap[itemIndex] = initialItemIndexMap[itemIndex]
+    app.itemIndexMap[itemIndex] = initialItemIndexMap[itemIndex]
   end
 end
 
+-- ========================
+-- Section: Additional merchant buttons
+-- ========================
+
 -- Equip item from merchant frame
-function merchantEquipHandler(i)
+function app:merchantEquipHandler(i)
   local itemFrame = _G["MerchantItem" .. i .. "ItemButton"]
-  local equipBtn = buttonBuilder({
+  local equipBtn = app:buttonBuilder({
     buttonName = "MerchantItem" .. i .. "EquipBtn",
     parent = itemFrame,
     text = "Equip",
@@ -345,8 +212,8 @@ function merchantEquipHandler(i)
     if button == "LeftButton" then
       local index = tonumber(self:GetName():match("%d+"))
       local sourceId = merchantIndex[index]
-      local source = getItemDetails(sourceId)
-      local slots = getItemSlot(source.itemEquipLoc)
+      local source = app:getItemDetails(sourceId)
+      local slots = app:getItemSlot(source.itemEquipLoc)
       if slots ~= 0 then
         local slotTarget = 0
         for _, slot in pairs(slots) do
@@ -359,6 +226,86 @@ function merchantEquipHandler(i)
   end)
 end
 
+-- IN Work
+-- Token display possible rewards class based
+function handleSourceMerchantToken(i)
+  local itemFrame = _G["MerchantItem" .. i]
+  local equipBtn = app:buttonBuilder({
+    buttonName = "MerchantItem" .. i .. "ShowSrcBtn",
+    parent = itemFrame,
+    text = "Source",
+    width = 50,
+    height = 22,
+    point = {
+      pos = "BOTTOMRIGHT",
+      x = 0,
+      y = 0,
+    }
+  })
+  equipBtn:HookScript("OnClick", function(self, button)
+    if button == "LeftButton" then
+      local index = tonumber(self:GetName():match("%d+"))
+      local sourceId = merchantIndex[index]
+      local source = app:getItemDetails(sourceId)
+      local tokenSrc = app.merchantTokenMap[source.itemId]
+      -- print(tokenSrc)
+      for _, item in pairs(tokenSrc) do
+        -- owned = isOwned, itemId = itemId
+        if (not item.owned) then
+          local details = app:getItemDetails(item.itemId) ---@type CH.ItemDetails
+          --print(details.link)
+          --print(item.owned)
+          --print(item.itemId)
+          --local details = getItemDetails(item.itemId)
+          --print(details)
+          --print(details.itemLink)
+          local t =
+          "|cffa335ee|Hitem:159300::168639::::::59:103::35:7:6536:6578:1543:4786:6439:6470:6514::::|h[Kula's Butchering Wristwraps]|h|r"
+          --sourceDetails:SetText(details.link)
+          app:createUserFrame(details.link)
+        end
+      end
+      app.merchantSourceItems:Show()
+      app.merchantFrameCost:SetPoint("TOPRIGHT", 280 * 2, 0)
+    end
+  end)
+end
+
+function app:createUserFrame(t)
+  local function createText(parent, w, h, x, y)
+    local t = parent:CreateFontString(nil, "OVERLAY", "GameFontHighlight")
+    t:SetSize(w, h)
+    t:SetPoint("TOPLEFT", x, y)
+    t:SetJustifyH("LEFT")
+    t:SetWordWrap(false)
+    return t
+  end
+
+  local userFrame = CreateFrame("Frame", nil, app.merchantSourceItems, "BackdropTemplate")
+  userFrame:SetClipsChildren(true)
+  userFrame:SetSize(492 + 24, 24)
+  userFrame:SetBackdrop({
+    bgFile = "Interface\\Buttons\\WHITE8x8"
+  })
+  userFrame:SetBackdropColor(0, 0, 0)
+  userFrame:SetPoint("TOPLEFT", 0, 0)
+  --userFrame:SetText(t)
+  local userName = createText(userFrame, 240, 26, 6, 1)
+  userName:SetText(t)
+end
+
+function toogleSourceTab()
+  if sourceHidden == true then
+    app.merchantFrameCost:SetPoint("TOPRIGHT", 280, 0)
+    app.merchantSourceItems:Hide()
+    sourceHidden = false
+  else
+    app.merchantSourceItems:Show()
+    app.merchantFrameCost:SetPoint("TOPRIGHT", 280 * 2, 0)
+    sourceHidden = true
+  end
+end
+
 -- ========================
 -- Section: Addon RegisterEvent
 -- ========================
@@ -368,12 +315,33 @@ CollectorHelper:RegisterEvent("TRANSMOG_COLLECTION_SOURCE_ADDED")
 CollectorHelper:RegisterEvent("GET_ITEM_INFO_RECEIVED")
 CollectorHelper:RegisterEvent("UNIT_SPELLCAST_SUCCEEDED")
 CollectorHelper:SetScript("OnEvent", function(self, event, arg1, arg2, arg3)
+  app:doSettings()
   if event == "ADDON_LOADED" and arg1 == "CollectorHelper" then
-    if settings == nil then
-      settings = settings or { hideMerchantOwned = true }
-    end
-    hooksecurefunc("MerchantFrame_Update", updateShop)
-    MerchantNextPageButton:HookScript("OnClick", updateShop)
-    MerchantPrevPageButton:HookScript("OnClick", updateShop)
+    hooksecurefunc("MerchantFrame_Update", function() app:updateShop() end)
+    MerchantNextPageButton:HookScript("OnClick", function() app:updateShop() end)
+    MerchantPrevPageButton:HookScript("OnClick", function() app:updateShop() end)
   end
 end)
+
+
+function app:doSettings()
+  if settings == nil then
+    settings = {
+      hideMerchantOwned = true,
+      showCostFrame = true,
+    }
+  end
+  app:MigrateSettings()
+
+  if settings.showCostFrame == false then
+      app.merchantFrameCost:Hide()
+  end
+end
+
+
+function app:MigrateSettings()
+  -- Check if settings table exists and if migration is needed
+  if settings.showCostFrame == nil then
+      settings.showCostFrame = true
+  end
+end